refactor: Map 객체 생성 함수 utils로 분리
- WinningStatistics 클래스 내부에서도 초기화 하는 로직에 재사용하기 위함
