+++ __tests__/WinningStatistics.test.js
@@ -1,50 +1,6 @@
-import { create } from "domain";
-import {
-  MATCH_KEY,
-  MIN_UNIT,
-  MATCH_PRIZE,
-} from "../src/constants/constants.js";
+import { MATCH_KEY, MIN_UNIT } from "../src/constants/constants.js";
 import WinningStatistics from "../src/domains/WinningStatistics.js";
-
-const createWinningStatisticsMap = (counts = {}) => {
-  return new Map([
-    [
-      MATCH_KEY.THREE,
-      {
-        count: counts[MATCH_KEY.THREE] ?? 0,
-        amount: MATCH_PRIZE[MATCH_KEY.THREE],
-      },
-    ],
-    [
-      MATCH_KEY.FOUR,
-      {
-        count: counts[MATCH_KEY.FOUR] ?? 0,
-        amount: MATCH_PRIZE[MATCH_KEY.FOUR],
-      },
-    ],
-    [
-      MATCH_KEY.FIVE,
-      {
-        count: counts[MATCH_KEY.FIVE] ?? 0,
-        amount: MATCH_PRIZE[MATCH_KEY.FIVE],
-      },
-    ],
-    [
-      MATCH_KEY.FIVE_AND_BONUS,
-      {
-        count: counts[MATCH_KEY.FIVE_AND_BONUS] ?? 0,
-        amount: MATCH_PRIZE[MATCH_KEY.FIVE_AND_BONUS],
-      },
-    ],
-    [
-      MATCH_KEY.SIX,
-      {
-        count: counts[MATCH_KEY.SIX] ?? 0,
-        amount: MATCH_PRIZE[MATCH_KEY.SIX],
-      },
-    ],
-  ]);
-};
+import createWinningStatisticsMap from "../src/utils/createWinningStatisticsMap.js";
 
 describe("WinningStatistics 클래스 테스트: 당첨 내역 통계 및 수익률 계산", () => {
   // given
