+++ src/domains/WinningStatistics.js
@@ -1,17 +1,10 @@
 import countMatchingNumbers from "../utils/countMatchingNumbers.js";
-import { MATCH_KEY, MATCH_PRIZE } from "../constants/constants.js";
+import { MATCH_KEY } from "../constants/constants.js";
+import createWinningStatisticsMap from "../utils/createWinningStatisticsMap.js";
+
 class WinningStatistics {
   #lottos = [];
-  #statistics = new Map([
-    [MATCH_KEY.THREE, { count: 0, amount: MATCH_PRIZE[MATCH_KEY.THREE] }],
-    [MATCH_KEY.FOUR, { count: 0, amount: MATCH_PRIZE[MATCH_KEY.FOUR] }],
-    [MATCH_KEY.FIVE, { count: 0, amount: MATCH_PRIZE[MATCH_KEY.FIVE] }],
-    [
-      MATCH_KEY.FIVE_AND_BONUS,
-      { count: 0, amount: MATCH_PRIZE[MATCH_KEY.FIVE_AND_BONUS] },
-    ],
-    [MATCH_KEY.SIX, { count: 0, amount: MATCH_PRIZE[MATCH_KEY.SIX] }],
-  ]);
+  #statistics = createWinningStatisticsMap();
 
   constructor(lottos) {
     this.#lottos = lottos;
